Rename EricaAuftragScheme -> EricaRequestSchema
