+++ erica/application/EricaAuftrag/EricaAuftragService.py
@@ -3,7 +3,7 @@ from uuid import UUID
 
 from opyoid import Injector, Module
 
-from erica.infrastructure.InfrastructureModule import InfrastructureModule
+from erica.infrastructure.infrastructure_module import InfrastructureModule
 from erica.infrastructure.rq.RqModule import RqModule
 from erica.infrastructure.sqlalchemy.repositories.erica_request_repository import EricaRequestRepository
